+++ src/connector/paginator/curl/mod.rs
@@ -33,7 +33,7 @@ impl PaginatorType {
         match self {
             PaginatorType::Offset(paginator) => {
                 let mut paginator = paginator.clone();
-                if paginator.count.is_none() {
+                if paginator.count.is_none() && connector.counter_type.is_some() {
                     paginator.count = Some(connector.clone().len().await?);
                 }