+++ src/updater/tera_helpers/function/mod.rs
@@ -1,0 +1,3 @@
+pub mod faker;
+pub mod object;
+pub mod string;