+++ src/updater/tera_helpers/mod.rs
@@ -1,3 +1,2 @@
 pub mod filters;
 pub mod function;
-pub mod faker;